First successful try

+++ distribution/bin_second_largest_values/bin_second_largest_values.cpp
@@ -3,23 +3,23 @@
 #include <algorithm>
 #include <set>
 
-std::vector<float> bin_second_largest_values(const std::vector<std::vector<float>>& prob_vectors, float bin_width) {
-    std::vector<float> second_largest_values;
+std::vector<double> bin_second_largest_values(const std::vector<std::vector<double>>& prob_vectors, double bin_width) {
+    std::vector<double> second_largest_values;
     for (const auto& vector : prob_vectors) {
-        std::vector<float> sorted_vector = vector;
+        std::vector<double> sorted_vector = vector;
         std::sort(sorted_vector.begin(), sorted_vector.end());
         second_largest_values.push_back(sorted_vector[sorted_vector.size() - 2]);
     }
 
-    std::set<float> binned_values_set;
+    std::set<double> binned_values_set;
     for (const auto& value : second_largest_values) {
-        float binned_value = (static_cast<int>(value / bin_width)) * bin_width;
+        double binned_value = (static_cast<int>(value / bin_width)) * bin_width;
         if (binned_value != 0) {
             binned_values_set.insert(binned_value);
         }
     }
 
-    std::vector<float> binned_values(binned_values_set.begin(), binned_values_set.end());
+    std::vector<double> binned_values(binned_values_set.begin(), binned_values_set.end());
     std::sort(binned_values.begin(), binned_values.end());
 
     return binned_values;
