+++ distribution/bin_second_largest_values/bin_second_largest_values.h
@@ -3,6 +3,6 @@
 
 #include <vector>
 
-std::vector<float> bin_second_largest_values(const std::vector<std::vector<float>>& prob_vectors, float bin_width);
+std::vector<double> bin_second_largest_values(const std::vector<std::vector<double>>& prob_vectors, double bin_width);
 
 #endif // BIN_SECOND_LARGEST_VALUES_H
