+++ distribution/find_closest_indices/example.cpp
@@ -3,12 +3,12 @@
 #include <vector>
 
 int main() {
-    std::vector<std::vector<float>> vecs = {
-        {4.0f, 2.0f, 5.0f},
-        {1.0f, 3.0f, 2.0f},
-        {8.0f, 7.0f, 6.0f}
+    std::vector<std::vector<double>> vecs = {
+        {4.0, 2.0, 5.0},
+        {1.0, 3.0, 2.0},
+        {8.0, 7.0, 6.0}
     };
-    std::vector<float> vals = {2.5f, 4.1f, 6.0f, 7.5f};
+    std::vector<double> vals = {2.5, 4.1, 6.0, 7.5};
 
     std::vector<int> indices = find_closest_indices(vecs, vals);
