+++ distribution/find_closest_indices/find_closest_indices.h
@@ -3,6 +3,6 @@
 
 #include <vector>
 
-std::vector<int> find_closest_indices(const std::vector<std::vector<float>>& vectors, const std::vector<float>& values);
+std::vector<int> find_closest_indices(const std::vector<std::vector<double>>& vectors, const std::vector<double>& values);
 
 #endif // FIND_CLOSEST_INDICES_H
