+++ distribution/quantile_1_minus_alpha/quantile_1_minus_alpha.cpp
@@ -2,16 +2,16 @@
 #include <vector>
 #include <algorithm>
 
-float quantile_1_minus_alpha(const std::vector<float>& values, const std::vector<float>& probabilities, float alpha) {
-    std::vector<std::pair<float, float>> value_prob_pairs;
+double quantile_1_minus_alpha(const std::vector<double>& values, const std::vector<double>& probabilities, double alpha) {
+    std::vector<std::pair<double, double>> value_prob_pairs;
     for (size_t i = 0; i < values.size(); ++i) {
         value_prob_pairs.emplace_back(values[i], probabilities[i]);
     }
 
     std::sort(value_prob_pairs.begin(), value_prob_pairs.end());
 
-    float cumulative_probability = 0.0f;
-    float quantile_threshold = 1.0f - alpha;
+    double cumulative_probability = 0.0;
+    double quantile_threshold = 1.0 - alpha;
 
     for (const auto& pair : value_prob_pairs) {
         cumulative_probability += pair.second;
