+++ utils/FilterVectors/FilterVectors.h
@@ -3,6 +3,6 @@
 
 #include <vector>
 
-std::vector<std::vector<float>> filter_vectors_by_max_value(const std::vector<std::vector<float>>& vectors, float threshold);
+std::vector<std::vector<double>> filter_vectors_by_max_value(const std::vector<std::vector<double>>& vectors, double threshold);
 
 #endif // FILTERVECTORS_H
