+++ utils/unique_vectors/unique_vectors.cpp
@@ -1,14 +1,13 @@
 #include "unique_vectors.h"
 #include <set>
 #include <vector>
-#include <algorithm>
 
-std::vector<std::vector<float>> unique_vectors(const std::vector<std::vector<float>>& vectors) {
-    std::set<std::set<float>> seen;
-    std::vector<std::vector<float>> unique_list;
+std::vector<std::vector<double>> unique_vectors(const std::vector<std::vector<double>>& vectors) {
+    std::set<std::set<double>> seen;
+    std::vector<std::vector<double>> unique_list;
 
     for (const auto& vector : vectors) {
-        std::set<float> vector_set(vector.begin(), vector.end());
+        std::set<double> vector_set(vector.begin(), vector.end());
         if (seen.find(vector_set) == seen.end()) {
             seen.insert(vector_set);
             unique_list.push_back(vector);
